% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backwards.rank.R
\name{BackwardsRank}
\alias{BackwardsRank}
\title{backwards rank}
\usage{
BackwardsRank(correlations, by.column = TRUE, similarity.measure = TRUE)
}
\arguments{
\item{correlations}{is the correlation-like (similarity) matrix to be processed it is supposed to be ranked; it is desribed by \code{similarity.measure} parameter}

\item{by.column}{- if \code{by.column==TRUE} (default) columns of returned object corresponds to tester (our) genes:  if \code{by.column==FALSE}, the rows corrdspond to tester genes}
}
\value{
\code{matrix} object if \code{correlations} is a matrix-like object; \code{data.table} if it is \code{data.table}; error otherwise; each column (or row if \code{by.column}, see \code{by.column} parameter description) of the object is ranks in columns of corelations matrix
}
\description{
The simplest function of the family, just rank data in columns and return
it in columns if by \code{by.column==TRUE}
}

