% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the.best.friends.r
\name{best.friends.of.features}
\alias{best.friends.of.features}
\title{best.friends.of.features}
\usage{
best.friends.of.features(relation, distance_like = FALSE)
}
\arguments{
\item{relation}{is the fetures*entities matrix of the relations between features and the entities}

\item{distance_like}{the default is \code{FALSE} and it shows that the relation values are not like distance, i.e. the better relation is shown by the lagrer value; is the relation is, on the contrary, distance-like, and 0 is the best, the value is \code{TRUE}.}
}
\value{
\code{data.frame} with 5 columns: feature index, friend entity index, uncorrected p-value for the pair, feature name, friend name
Best friend has the highest order, the worst has the lowest
}
\description{
We have someones to be a friend (entity) as columns and we have what is to be friends of (features) as rows.
}
