% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.friends.of.r, R/the.best.friends.r
\docType{package}
\name{best.friends.of}
\alias{best.friends.of}
\title{best.friends.of: A package that provide the best-friend ranking}
\usage{
best.friends.of(relation, distance_like = FALSE)
}
\arguments{
\item{relation}{is the fetures*entities matrix of the relations between features ans entities}
}
\value{
\code{data.frame} with 3 columns: feature index, friend entity index, uncorrected p-value for the pair
Best friend has the highest order, the worst has the lowest
}
\description{
Suppose We have some similarity of interaction measure for pairs of instances, e.q. genes and regulators.
We want to know what genes are the specifically best friends for selected regulators and vice versa.
}
\details{
Main function:
}
\section{best.friends.of functions}{

best.friends.of takes the matrix, provide the potential marker - best frind index - p-value table
}

