% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the.best.friends.r
\name{best.friends.test}
\alias{best.friends.test}
\title{best.friends.test}
\usage{
best.friends.test(relation, distance_like = FALSE)
}
\arguments{
\item{relation}{is the elements*communities matrix of the relations between features and the entities}

\item{distance_like}{the default is \code{FALSE} and it shows that the relation values are not like distance, i.e. the better relation is shown by the lagrer value; if the relation is, on the contrary, distance-like, and 0 is the best, the value is \code{TRUE}.}
}
\value{
\code{data.frame} with 5 columns: element index, the index of the community that is a putative best friend of the element, uncorrected p-value for the pair, element name, friend name. The small (after correctio we are to do) p-value indicates that the community is really the best friend of the element
}
\description{
We have what can to be friends (communities) as columns and we have what they can befriends of (features) as rows.
}
