% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwards.rank.R
\name{DistanceByBackwardsRank}
\alias{DistanceByBackwardsRank}
\title{create distance matrix based on the rank-reverse-rank
best friends first}
\usage{
DistanceByBackwardsRank(correlations, similarity.measure = TRUE)
}
\arguments{
\item{similarity.measure;}{if \code{TRUE}, the measure in the matrix is similarity (the larger it is, the closer are the genes); if it is \code{FALSE}, it is distance}
}
\value{
\code{matrix} object if \code{correlations} is a matrix-like object; \code{data.table} if it is \code{data.table}; error otherwise; the return is new distance matrix
}
\description{
create distance matrix based on the rank-reverse-rank
best friends first
}
