% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the.best.friends.r
\name{friends.of.features}
\alias{friends.of.features}
\title{friends.of.features}
\usage{
friends.of.features(relation, distance_like = FALSE, friends_num = -1)
}
\arguments{
\item{relation}{is the fetures*entities matrix of the relations between features and the entities}

\item{distance_like}{the default is \code{FALSE} and it shows that the relation values are not like distance, i.e. the better relation is shown by the lagrer value; is the relation is, on the contrary, distance-like, and 0 is the best, the value is \code{TRUE}.}

\item{friends_num}{number of entities we consider for each feature; the default -1 means all; if friends_num is 1,}
}
\value{
\code{data.frame} with 4 columns: friend entity index, uncorrected p-value for the pair, feature name, friend name
Best friend has the highest order, the worst has the lowest
}
\description{
We have someones to be a friend (entity) as columns and we have what is to be friends of (features) as rows.
}
