% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the.best.friends.r
\name{OrderByBestFriendOf}
\alias{OrderByBestFriendOf}
\title{OrderByBestFriendOf
We have someones to be friended by as columns and we have the possible frieds as rows. 
The function outputs a matrix where in each column, the list of the numbers of the rows are ordered by how this row is specifically friendly to this column. Create gene number lists that are orders of genes in ranking by backwards-rank of the genenes correlations matrix.
\code{my.row.names[number list]} will give the sorted list of gene nemes}
\usage{
OrderByBestFriendOf(friendship)
}
\arguments{
\item{friendship}{is the freindship (e.g.similarity) matrix to be processed; the someones to be friended by are columns; possible friends are in rows}
}
\value{
\code{matrix} object if \code{frriendship} is a matrix-like object; \code{data.table} if it is \code{data.table}; error otherwise; each column of the object is a list of orders of genes as sorted by best friends rank in the column (if \code{by.column})
}
\description{
OrderByBestFriendOf
We have someones to be friended by as columns and we have the possible frieds as rows. 
The function outputs a matrix where in each column, the list of the numbers of the rows are ordered by how this row is specifically friendly to this column. Create gene number lists that are orders of genes in ranking by backwards-rank of the genenes correlations matrix.
\code{my.row.names[number list]} will give the sorted list of gene nemes
}
