% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rank_diff_and_p}
\alias{rank_diff_and_p}
\title{Identifying the putative friend entity for the vector of ranks of the feature in different entities.}
\usage{
rank_diff_and_p(x)
}
\arguments{
\item{x}{the list of ranks normalised to 0..1 -- in null, they are independently distributed}
}
\value{
a vector of: index of the best, the best rank-the next rank and the p-value (it is the difference^n)
}
\description{
Identifying the putative friend entity for the vector of ranks of the feature in different entities.
}
