% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rank_diff_and_p_for_the_best}
\alias{rank_diff_and_p_for_the_best}
\title{Estimating the p-value for the putative entity that is the best friend of a feature from a vestor of normalised ranks of the feature for different entities.
The putative friend is that with the the best (mininal) rank in the list.
The corresponding p-value is the difference of the next and the best to power of the size of the list of entities
next here means the next by rank rather than by index in x}
\usage{
rank_diff_and_p_for_the_best(x)
}
\arguments{
\item{x}{the list of ranks of the feature in different entities normalised to 0..1 -- in null, they are independently distributed}
}
\value{
a vector of: index of the best and the p-value (it is the the_next_value-the_best_value^n)
}
\description{
Estimating the p-value for the putative entity that is the best friend of a feature from a vestor of normalised ranks of the feature for different entities.
The putative friend is that with the the best (mininal) rank in the list.
The corresponding p-value is the difference of the next and the best to power of the size of the list of entities
next here means the next by rank rather than by index in x
}
