% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwards.rank.R
\name{RankByBackwardsRank}
\alias{RankByBackwardsRank}
\title{Rank by backwards rank
Creates the matrix of rankes of genes by the backwards ranks in correlation matrix}
\usage{
RankByBackwardsRank(correlations, by.column = TRUE,
  similarity.measure = TRUE)
}
\arguments{
\item{correlations}{is the correlation-like (similarity) matrix to be processed it is supposed to be ranked; it is desribed by \code{similarity.measure} parameter}

\item{by.column}{- if \code{by.column==TRUE} (default) columns of returned object corresponds to tester (our) genes:  if \code{by.column==FALSE}, the rows corrdspond to tester genes}
}
\value{
\code{matrix} object if \code{correlations} is a matrix-like object; \code{data.table} if it is \code{data.table}; error otherwise; each column (or row, see \code{by.column} parameter description) of the object represet ranks of genes by backwards-rank relation to the tester gene tester gene is column if \code{by.column}; the possible friend of tester gene is row; the lower is the rank, the better is the friend
}
\description{
Rank by backwards rank
Creates the matrix of rankes of genes by the backwards ranks in correlation matrix
}
