% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the.best.friends.r
\name{RankByBestFriendOf}
\alias{RankByBestFriendOf}
\title{RankByBestFriendOf
We have someones to be friended by as columns and we have the possible frieds as rows.
The function outputs a matrix where in each column and row there is a value of how the raw specifically friendly to this column.
\code{my.row.names[number list]} will give the sorted list of gene nemes
Best friend has the lowest order.}
\usage{
RankByBestFriendOf(friendship)
}
\arguments{
\item{friendship}{is the freindship (e.g.similarity) matrix to be processed; the someones to be friended by are columns; possible friends are in rows}
}
\value{
\code{matrix} object if \code{frriendship} is a matrix-like object; \code{data.table} if it is \code{data.table}; error otherwise; each column (or row if \code{by.column}, see \code{by.column} parameter description) of the object is is a matrix of best frindship ranks of rows for the column. Provide usual increasing ranks.
Best friend has the highest order, the worst has the lowest
export
}
\description{
RankByBestFriendOf
We have someones to be friended by as columns and we have the possible frieds as rows.
The function outputs a matrix where in each column and row there is a value of how the raw specifically friendly to this column.
\code{my.row.names[number list]} will give the sorted list of gene nemes
Best friend has the lowest order.
}
