% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backwards.rank.R
\name{sort.by.backwards.rank}
\alias{sort.by.backwards.rank}
\title{sort.by.backwards.rank}
\usage{
\method{sort}{by.backwards.rank}(correlations, by.column = TRUE,
  similarity.measure = TRUE)
}
\arguments{
\item{correlations}{is the correlation-like (similarity) matrix to be processed it is supposed to be ranked; it is desribed by \code{similarity.measure} parameter}

\item{by.column}{- if \code{by.column==TRUE} (default) columns of returned object corresponds to tester (our) genes:  if \code{by.column==FALSE}, the rows corrdspond to tester genes}

\item{similarity.measure;}{if \code{TRUE}, the \code{correlations} matrix is similarity (the larger it is, the closer are the genes); if it is \code{FALSE}, it is distance}
}
\value{
\code{matrix} object if \code{correlations} is a matrix-like object; \code{data.table} if it is \code{data.table}; error otherwise; each column (or row, see \code{by.column} parameter description) of the object is is a list of genes sorted by backwards rank by relation to the tester gene
}
\description{
create gene name lists that are sorted by backwards-rank in the correlations matrix
best friends first
}

