% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.friends.of.r
\docType{package}
\name{best.friends.of}
\alias{best.friends.of}
\title{best.friends.of: A package that provide the best-friend ranking}
\description{
Suppose We have some similarity of interaction measure for pairs of instances, e.q. genes and regulators.
We want to know what genes are the specifically best friends for selected regulators and vice versa.
}
\details{
Main function:
}
\section{best.friends.of functions}{

RankByBestFriendOf takes the matrix, provide the ranks for best-friend ranking 
OrderByBestFriendOf takes the matrix, provide the orders by best-friend ranking
}

