% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the.best.friends.r
\name{friends.test}
\alias{friends.test}
\title{friends.test}
\usage{
friends.test(relation, distance_like = FALSE, friends.number = -1)
}
\arguments{
\item{relation}{is the elements*communities matrix of the relations between features and the entities}

\item{distance_like}{the default is \code{FALSE} and it shows that the relation values are not like distance, i.e. the better relation is shown by the lagrer value; if the relation is, on the contrary, distance-like, and 0 is the best, the value is \code{TRUE}.}

\item{friends.number}{number of entities we consider for each feature; the default -1 means all; if friends.number is 1,
the call do essentialy the same as the best.friends.test call}
}
\value{
\code{data.frame} with 4 columns: friend entity index, uncorrected p-value for the pair, feature name, friend name
Best friend has the highest order, the worst has the lowest
}
\description{
We have what can to be friends (communities) as columns and we have what they can befriends of (features) as rows.
}
